feat: 회원가입 후 스플래시 적용
